Test/requirements copy.txt
qiskit==1.2.4
qiskit-aer-gpu==0.17.2
qiskit-optimization==0.7.0
qiskit-ibm-runtime==0.44.0

annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.1
attrs==26.1.0
cached-property==1.5.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cirq-core==1.6.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
deprecation==2.1.0
dill==0.4.0
docplex==2.32.262
duet==0.2.9
executing==2.2.1
fonttools==4.61.1
h5py==3.15.0
hashable-list==0.2.0
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.72.0
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.6.1
numpy==1.26.4
openfermion==1.7.1
openfermionpyscf==0.5
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
pip==25.3
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.2.1
ptyprocess==0.7.0
PubChemPy==1.0.5
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.3.1
pyscf==2.11.0
pyspnego==0.12.0
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
qiskit-algorithms==0.3.1
qiskit-nature==0.7.2
requests==2.32.5
requests_ntlm==1.3.0
rustworkx==0.17.1
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
stack_data==0.6.3
stevedore==5.6.0
symengine==0.13.0
sympy==1.14.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.2.14
wheel==0.45.1
